rework parallelism in collectsnmp module
Move away from Pool to a mix of Queue and Process.  This creates a task
queue that is shared among the workers.  This also pushes most of the
global methods to be methods in CollectSnmpWorker.  The small methods at
the end are static methods just because they do trivial work and there's
no state.

This also implements the option to limit work to modules that collect
measurement data instead of just informational data.